+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `ovni_version_get()` function.
 - Add the `ovniver` program to report the libovni version and commit.
 - Add nOS-V API subsystem events for `nosv_create()` and `nosv_destroy()`.
+- Add TAMPI model with `T` code.
+- Add subsytem events and cfgs for TAMPI model.
 
 ## [1.2.2] - 2022-07-26